pyproject.toml
[project]
name = "Projet2A_2024"
version = "0.1.0"
description = "Teacher's version of the Projet2A 2024"
authors = [
    {name = "ClémentValot", email = "clement.valot.pro@gmail.com"},
]
dependencies = [
    "fastapi>=0.112.1",
    "uvicorn>=0.30.6",
    "pydantic>=2.8.2",
    "PyJWT>=2.9.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "apscheduler>=3.11.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pytest.ini_options]
addopts = "-ra -q"
pythonpath = [ "."] 
testpaths = [
    "tests",
]


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
test = [
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
    "freezegun>=1.5.1",
]
lint = [
    "ruff>=0.6.0",
]
typing = [
    "mypy>=1.11.2",
    "types-requests>=2.32.0.20240905",
    "types-psycopg2>=2.9.21.20240819",
]

[tool.pdm.scripts]
lint = "ruff check"
format = "ruff format"
test = "pytest tests --cov=src"
start = "pdm run __main__.py"
typecheck = "pdm run mypy ."

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "B", "I", "A", "C90", "PERF", "TCH", "TID"]

[tool.mypy]
plugins = ["pydantic.mypy"]
